backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
email-validator==2.1.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosmtplib==3.0.1
slowapi==0.1.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
redis==5.0.1
# Phase 4 document processing
PyMuPDF==1.23.8
python-docx==1.1.0
sentence-transformers==2.2.2
chromadb==0.4.18
tiktoken==0.5.2
aiofiles==23.2.1
# RAG improvements
openai>=1.0.0